requirements.txt
cumm_cu120==0.4.11
Cython==3.0.12
d3d==0.1.0rc0
denoising_diffusion_pytorch==2.1.1
diffusers==0.33.1
efficientnet_pytorch==0.7.0
einops==0.8.1
h5py==3.13.0
icecream==2.1.4
imageio==2.37.0
matplotlib==3.10.8
MultiScaleDeformableAttention==1.0
numpy==1.24.4
open3d==0.19.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==12.1.0
pyquaternion==0.9.9
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==75.8.0
Shapely==2.1.2
skimage==0.0
spconv_cu120==2.3.6
tensorboard==2.19.0
tensorboardX==2.6.2.2
tensorboardX==2.6.4
termcolor==3.3.0
timm==1.0.23
torch==2.1.0+cu121
torch_scatter==2.1.2
torchvision==0.16.0+cu121
tqdm==4.67.1
